refactor(telegram): import TelegramSender from sender, drop handlers re-export
Removes the ``TelegramSender as TelegramSender`` re-export shim from
``handlers.py`` and updates every caller to import directly from
``app.integrations.telegram.sender`` — the real owner. No more
backwards-compatibility surface; each module imports the type from
where it lives.

Touched callers:
- ``bot.py``
- ``model_picker_runtime.py``
- ``tests/test_channels_api.py``
- ``tests/test_telegram_channel.py``
- ``tests/test_telegram_dev_admin.py``
- ``tests/test_telegram_model_picker.py``
